pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "gnnfairviz"
authors = [
    {name = "Xinwu Ye", email = "yexw0403@foxmail.com"},
]
description = "Official implementation of GNNFairViz"
readme = "README.md"
requires-python = ">=3.8,<=3.11"
keywords = ["fairness", "visual analysis", "graph neural networks", "Jupyter notebook"]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
]
# dependencies = [
#     "numpy==1.26.2",
#     "holoviews==1.16.2",
#     "torch==2.0.1",
#     "dgl==1.1.1+cu113",
#     "pandas==1.5.3",
#     "panel==1.2.0",
#     "shapely==2.0.1",
#     "scikit-learn==1.3.0",
#     "umap-learn==0.5.5",
#     "scipy==1.11.4",
#     "tqdm==4.65.0",
#     "bokeh==3.1.1",
#     "matplotlib==3.7.1",
# ]
dependencies = [
    "numpy>=1.26.2",
    "holoviews>=1.16.2",
    "torch>=2.0.1",
    "dgl>=1.1.1",
    "pandas>=1.5.3",
    "panel>=1.2.0",
    # "shapely>=2.0.1",
    "shapely",
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.5",
    "scipy>=1.11.4",
    "tqdm>=4.65.0",
    "bokeh>=3.1.1",
    "matplotlib>=3.7.1",
]

dynamic = ["version"]

[tool.setuptools]
packages = ["gnnfairviz"]

# ... other project metadata fields as listed in:
#     https://packaging.python.org/en/latest/guides/writing-pyproject-toml/